pyproject.toml
[project]
name = "parabola2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "reportlab>=4.4.4",
    "ruamel-yaml>=0.18.15",
]
